# Tail call VM