--- a/Zend/zend_portability.h
+++ b/Zend/zend_portability.h
@@ -313,6 +313,18 @@ char *alloca();
 # define ZEND_FASTCALL
 #endif
 
+#if __has_attribute(preserve_none) && !defined(__SANITIZE_ADDRESS__)
+#  define HAVE_PRESERVE_NONE
+#  define ZEND_PRESERVE_NONE __attribute__((preserve_none))
+#else
+#  define ZEND_PRESERVE_NONE
+#endif
+
+#if __has_attribute(musttail)
+#  define HAVE_MUSTTAIL
+#  define ZEND_MUSTTAIL __attribute__((musttail))
+#endif
+
 #if (defined(__GNUC__) && __GNUC__ >= 3 && !defined(__INTEL_COMPILER) && !defined(__APPLE__) && !defined(__hpux) && !defined(_AIX) && !defined(__osf__)) || __has_attribute(noreturn)
 # define HAVE_NORETURN
 # define ZEND_NORETURN __attribute__((noreturn))