… Zend/zend_vm_execute.h
… Zend/zend_vm_gen.php