@arnaud-lb 2025-02-18
Yes this does require some changes to the JIT to accommodate for the new opcode handler signature and how FP/IP are passed around. I plan to implement them unless there are major issues with the current approach.

The fact that `preserve_none` is an unstable ABI will complicate things a bit. Some possible solutions I have in mind are:
- Generate wrappers with a stable ABI for each opcode handler, and use that in opcache/jit
- Enforce that opcache must be compiled with the same clang version as the php binary
- Move opcache to Zend/ and embed it in the php binary

The second one seems reasonable to me.