@cmb69 2025-02-18
FWIW: [feature request to support guaranteed tail calls for MSVC](https://developercommunity.visualstudio.com/t/Please-support-guaranteed-tail-call-attr/10091382?).