@dstogov 2025-02-19
BTW LLVM/CLANG should support *local* register variables. So maybe GOTO and HYBRID VMs may be adopted.